requirements.txt
boto3~=1.18.42
botocore~=1.21.42
setuptools~=57.4.0
AWSIoTPythonSDK~=1.4.9
-e ./terraform-modules/layers/videostreams
-e ./terraform-modules/layers/lambda2kinesis
-e ./terraform-modules/layers/video_stream_data_archiver
requests~=2.27.1
pyathena~=2.1.0
pyspark~=3.2.1
pymongo~=4.0.2
python-dotenv~=1.0.0
pytest~=7.3.2
boto3_type_annotations~=0.3.1
pyorc~=0.9.0
moto~=5.0.28
